`object`:add|implement
