+++ sources/object.m
@@ -1,0 +1,9 @@
+#include <object.h>
+
+#include <mesh/mesh.h>
+
+void object_destroy(
+  struct object* object
+) {
+  mesh_destroy(&object->mesh);
+}
